move core module to the separate package
