+++ packages/skale-core/pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "skale.py-core"
+version = "0.1.0"
+description = "SKALE core settings and types"
+readme = "README.md"
+requires-python = ">=3.11,<4"
+license = { text = "AGPL-3.0-or-later" }
+authors = [{ name = "SKALE Labs", email = "support@skalelabs.com" }]
+keywords = ["skale"]
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "Intended Audience :: Developers",
+  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
+  "Natural Language :: English",
+  "Programming Language :: Python :: 3.11",
+]
+dependencies = [
+  "pydantic-settings>=2.12.0",
+  "tomli-w>=1.2.0",
+  "typing-extensions==4.15.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/skalenetwork/skale.py"
+
+[tool.hatch.build.targets.wheel]
+packages = ["skale_core"]
+
+[tool.ruff]
+line-length = 100
+target-version = "py314"
+
+[tool.ruff.lint]
+select = ["E", "W", "F", "I"]
+
+[tool.ruff.format]
+quote-style = "single"
