+++ packages/skale-core/skale_core/constants.py
@@ -17,6 +17,4 @@
 #   You should have received a copy of the GNU Affero General Public License
 #   along with SKALE.py.  If not, see <https://www.gnu.org/licenses/>.
 
-from skale_core.types import EnvType, ImageType, NodeMode, NodeType, SchainName
-
-__all__ = ['EnvType', 'ImageType', 'NodeMode', 'NodeType', 'SchainName']
+NESTED_DELIMITER = '__'
