+++ packages/skale-core/skale_core/helper.py
@@ -17,6 +17,21 @@
 #   You should have received a copy of the GNU Affero General Public License
 #   along with SKALE.py.  If not, see <https://www.gnu.org/licenses/>.
 
-from skale_core.types import EnvType, ImageType, NodeMode, NodeType, SchainName
+import os
+import tempfile
+from pathlib import Path
 
-__all__ = ['EnvType', 'ImageType', 'NodeMode', 'NodeType', 'SchainName']
+
+def _atomic_write_text(path: Path, text: str, mode: int = 0o600) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    fd, tmp_name = tempfile.mkstemp(prefix=path.name + '.', dir=str(path.parent))
+    try:
+        with os.fdopen(fd, 'w', encoding='utf-8') as f:
+            f.write(text)
+            f.flush()
+            os.fsync(f.fileno())
+        os.chmod(tmp_name, mode)
+        os.replace(tmp_name, path)
+    finally:
+        if Path(tmp_name).is_file():
+            Path(tmp_name).unlink()
